backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cfgv==3.5.0
click==8.3.0
distlib==0.4.0
fastapi==0.121.1
filelock==3.20.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
nodeenv==1.9.1
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
pre-commit==4.5.0
pydantic==2.12.4
pydantic-core==2.41.5
pygments==2.19.2
pytest==9.0.0
pyyaml==6.0.3
rich==14.2.0
ruff==0.14.7
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.3
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
virtualenv==20.35.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
python-dotenv==1.0.1